fix: show social buttons on signup page
